add new features

+++ src/Controller/TaskController.php
@@ -4,6 +4,7 @@ namespace App\Controller;
 
 use App\Entity\Task;
 use App\Form\TaskType;
+use App\Repository\TaskRepository;
 use Doctrine\ORM\EntityManagerInterface;
 use Symfony\Component\HttpFoundation\Request;
 use Symfony\Component\HttpFoundation\Response;
@@ -13,9 +14,9 @@ use Symfony\Bundle\FrameworkBundle\Controller\AbstractController;
 class TaskController extends AbstractController
 {
     #[Route('/tasks', name: 'task_list')]
-    public function listAction(EntityManagerInterface $em):response
+    public function listAction(TaskRepository $taskRepository):response
     {
-        $tasks = $em->getRepository(Task::class)->findAll();
+        $tasks = $taskRepository->findTasksForUser($this->getUser());
 
         return $this->render('task/list.html.twig', ['tasks' => $tasks]);
     }
@@ -24,6 +25,7 @@ class TaskController extends AbstractController
     public function createAction(Request $request, EntityManagerInterface $em): Response
     {
         $task = new Task();
+        $task->setAuthor($this->getUser());
         $form = $this->createForm(TaskType::class, $task);
 
         $form->handleRequest($request);
@@ -38,12 +40,19 @@ class TaskController extends AbstractController
             return $this->redirectToRoute('task_list');
         }
 
-        return $this->render('task/create.html.twig', ['form' => $form->createView()]);
+        return $this->render('task/create.html.twig', [
+            'form' => $form->createView()
+        ]);
     }
 
     #[Route('/tasks/{id}/edit', name: 'task_edit')]
     public function editAction(Task $task, Request $request, EntityManagerInterface $em): Response
     {
+        // Vérifie que l'utilisateur actuel est bien l'auteur de la tâche
+        if ($task->getAuthor() !== $this->getUser()) {
+            $this->addFlash('danger', "Vous n'avez pas le droit de modifier cette tâche.");
+        }
+
         $form = $this->createForm(TaskType::class, $task);
 
         $form->handleRequest($request);
@@ -65,6 +74,11 @@ class TaskController extends AbstractController
     #[Route('/tasks/{id}/toggle', name: 'task_toggle')]
     public function toggleTaskAction(Task $task, EntityManagerInterface $em): Response
     {
+        if ($task->getAuthor() !== $this->getUser()) {
+            $this->addFlash('danger', 'Vous ne pouvez pas modifier cette tâche.');
+            return $this->redirectToRoute('task_list');
+        }
+
         $task->toggle(!$task->isDone());
         $em->flush();
 
@@ -76,6 +90,15 @@ class TaskController extends AbstractController
     #[Route('/tasks/{id}/delete', name: 'task_delete')]
     public function deleteTaskAction(Task $task, EntityManagerInterface $em): Response
     {
+        $isAdmin = $this->isGranted('ROLE_ADMIN');
+        $isAuthor = $task->getAuthor() === $this->getUser();
+        $isAnonymousTask = $task->getAuthor()?->getUsername() === 'anonyme';
+
+        if (!$isAuthor && !($isAnonymousTask && $isAdmin)) {
+            $this->addFlash('danger', 'Vous ne pouvez pas supprimer cette tâche.');
+            return $this->redirectToRoute('task_list');
+        }
+        
         $em->remove($task);
         $em->flush();
